+++ apps/cms-v2/src/modules/channel-provider-connection/channel-provider-connections-section-header.tsx
@@ -1,0 +1,7 @@
+import { Text } from "@saleor/macaw-ui/next";
+
+export const ChanelProviderConnectionsSectionHeader = () => (
+  <Text marginBottom={4} as="h2" variant="heading">
+    Channels Connections
+  </Text>
+);
