+++ apps/cms-v2/src/modules/configuration/schemas/root-config.schema.ts
@@ -1,37 +1,44 @@
 import { z } from "zod";
-import { ChannelProviderConnectionConfigSchema } from "./channel-provider-connection.schema";
-import { ContentfulProviderSchema } from "./contentful-provider.schema";
-import { DatocmsProviderSchema } from "./datocms-provider.schema";
+import { ChannelProviderConnectionConfig } from "./channel-provider-connection.schema";
+import { ContentfulProviderConfig } from "./contentful-provider.schema";
+import { DatocmsProviderConfig } from "./datocms-provider.schema";
 import { StrapiProviderConfig } from "./strapi-provider.schema";
 
-export const AnyProviderConfigSchema = z.union([
+// todo move to shared bootstrap?
+export namespace ProvidersConfig {
+  export const AnyProviderConfigSchema = z.union([
+    /**
+     * Add more for each provider
+     */
+    ContentfulProviderConfig.Schema.Full,
+    DatocmsProviderConfig.Schema.Full,
+    StrapiProviderConfig.Schema.Full,
+  ]);
+
+  export const AnyProvidersListSchema = z.array(AnyProviderConfigSchema);
+
+  export const AnyProvidersInput = z.union([
+    ContentfulProviderConfig.Schema.Input,
+    DatocmsProviderConfig.Schema.Input,
+    StrapiProviderConfig.Schema.Input,
+  ]);
+
+  export type AnyProviderConfigSchemaType = z.infer<typeof AnyProviderConfigSchema>;
+  export type AnyProviderInputSchemaType = z.infer<typeof AnyProvidersInput>;
+}
+
+export namespace RootConfig {
   /**
-   * Add more for each provider
+   * Store entire app config in single file
+   * - Only one request
+   * - Always transactional
    */
-  ContentfulProviderSchema.Config,
-  DatocmsProviderSchema.Config,
-  StrapiProviderConfig.Schema.Full,
-]);
+  export const Schema = z.object({
+    providers: ProvidersConfig.AnyProvidersListSchema,
+    connections: z.array(ChannelProviderConnectionConfig.Schema.Full),
+  });
 
-const AnyProvidersListSchema = z.array(AnyProviderConfigSchema);
+  export type Shape = z.infer<typeof Schema>;
 
-export const AnyProvidersInput = z.union([
-  ContentfulProviderSchema.ConfigInput,
-  DatocmsProviderSchema.ConfigInput,
-  StrapiProviderConfig.Schema.Input,
-]);
-
-/**
- * Store entire app config in single file
- * - Only one request
- * - Always transactional
- */
-export const RootConfigSchema = z.object({
-  providers: AnyProvidersListSchema,
-  connections: z.array(ChannelProviderConnectionConfigSchema.Connection),
-});
-
-export type RootConfigSchemaType = z.infer<typeof RootConfigSchema>;
-
-export type AnyProviderConfigSchemaType = z.infer<typeof AnyProviderConfigSchema>;
-export type AnyProviderInputSchemaType = z.infer<typeof AnyProvidersInput>;
+  export type RootConfigSchemaType = z.infer<typeof Schema>;
+}
